ref(resolvers): move resolvers to their own folder
