+++ src/app/app.module.ts
@@ -18,10 +18,10 @@ import { ProjectComponent } from './project/project.component';
 import { BuildStatusBadgeComponent } from './build-status-badge/build-status-badge.component';
 import { ApiProjectService } from './services/project/api-project.service';
 import { BuildService } from './services/build/build.service';
-import { ProjectResolver } from './services/project.resolver';
-import { BuildsResolver } from './services/builds.resolver';
-import { ProjectsBuildResolver } from './services/projects-build.resolver';
-import { BuildResolver } from './services/build.resolver';
+import { ProjectResolver } from './services/resolvers/project.resolver';
+import { BuildsResolver } from './services/resolvers/builds.resolver';
+import { ProjectsBuildResolver } from './services/resolvers/projects-build.resolver';
+import { BuildResolver } from './services/resolvers/build.resolver';
 
 @NgModule({
   declarations: [
