+++ src/app/services/resolvers/builds.resolver.ts
@@ -1,15 +1,15 @@
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 
-import { BuildService } from './build/build.service';
-import { Build } from '../models/Build';
+import { BuildService } from '../build/build.service';
+import { Build } from '../../models/Build';
 
 @Injectable()
-export class BuildResolver implements Resolve<Build> {
+export class BuildsResolver implements Resolve<Build[]> {
 
   constructor(private buildService: BuildService) {}
 
   resolve(route: ActivatedRouteSnapshot) {
-    return this.buildService.getBuild(route.paramMap.get('id'));
+    return this.buildService.getBuilds(route.paramMap.get('id'));
   }
 }
