+++ src/app/services/resolvers/project.resolver.ts
@@ -3,8 +3,8 @@ import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/catch';
 
-import { ProjectService } from './project/project.service';
-import { Project } from '../models/Project';
+import { ProjectService } from '../project/project.service';
+import { Project } from '../../models/Project';
 
 @Injectable()
 export class ProjectResolver implements Resolve<Project> {
