+++ src/app/services/resolvers/projects-build.resolver.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 
-import { ProjectService } from './project/project.service';
-import { ProjectsBuild } from '../models/ProjectsBuild';
+import { ProjectService } from '../project/project.service';
+import { ProjectsBuild } from '../../models/ProjectsBuild';
 
 @Injectable()
 export class ProjectsBuildResolver implements Resolve<ProjectsBuild[]> {
